+++ apps/frontend/src/app/globals.module.css
@@ -5,10 +5,6 @@
 }
 
 .navbar {
-  margin: 0;
-  padding: 0;
-  top: 0;
-  left: 0;
   width: 100%;
   height: 60px;
   background-color: #7eb397;
@@ -22,15 +18,4 @@
   font-weight: 500;
   margin-left: 10px;
   margin-right: 10px;
-}
-
-.links-container {
-  margin-right: 10px;
-}
-
-.link a {
-  text-decoration: none;
-  color: #141414;
-  font-weight: 700;
-  font-size: 30px;
 }
