+++ apps/frontend/src/app/layout.tsx
@@ -4,8 +4,6 @@ import type { Metadata } from 'next';
 
 import Navbar from '@/components/navbar';
 
-/*const inter = Inter({ subsets: ['latin'] });*/
-
 export const metadata: Metadata = {
   title: 'ProgramSCH',
   description: 'created by KirDev',
@@ -18,7 +16,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang='hu'>
-      <body className='m-0'>
+      <body>
         <Navbar />
         {children}
       </body>
